Let Android choose the filename
This avoids fastFindFile, which is really not all the fast.
It may be possible to guess the documentId and optimize it, but
it wouldn't work on some exotic saving locations.

Better to just hand off the basename to Android, it will pick a new
name if there is a collision. So in the end it will do the same thing.

Hopefully this is faster.