Fix state machine mistake in PriorityQueue
reap could occur while people were still waiting for the lock on an queue. So then the add is called on reaped queue, and end called on fresh queue. Leading to throw on the end.

Now just reap every few minutes, it's a slow memory leak in any case. Only one object per host. Probably not that many hosts.